requirements.txt
cryptography
pyperclip
pyqt5-tools